file structure rework
